# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Diary Service
    description: Service for Diary
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /signin:
        post:
            tags:
                - diary
            summary: Signin diary
            description: Creates a valid API token
            operationId: diary#Signin
            parameters:
                - name: SigninRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/DiarySigninRequestBody'
                    required:
                        - email
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
    /signup:
        post:
            tags:
                - diary
            summary: UserSignup diary
            operationId: diary#UserSignup
            parameters:
                - name: UserSignupRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/DiaryUserSignupRequestBody'
                    required:
                        - name
                        - email
            responses:
                "201":
                    description: Created response.
            schemes:
                - http
    /users/{user_name}/diaries:
        post:
            tags:
                - diary
            summary: CreateDiary diary
            description: Creates a diary
            operationId: diary#CreateDiary
            parameters:
                - name: user_name
                  in: path
                  description: User name
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  description: API key used to perform authorization
                  required: false
                  type: string
                - name: CreateDiaryRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/DiaryCreateDiaryRequestBody'
                    required:
                        - title
            responses:
                "201":
                    description: Created response.
            schemes:
                - http
            security:
                - api_key_header_Authorization: []
definitions:
    DiaryCreateDiaryRequestBody:
        title: DiaryCreateDiaryRequestBody
        type: object
        properties:
            title:
                type: string
                description: Diary title
                example: Culpa pariatur in assumenda voluptatem rerum.
        example:
            title: Ut magnam vel ut repellat asperiores.
        required:
            - title
    DiarySigninRequestBody:
        title: DiarySigninRequestBody
        type: object
        properties:
            email:
                type: string
                description: User email
                example: Tempore incidunt.
        example:
            email: Similique sint cupiditate id et enim.
        required:
            - email
    DiaryUserSignupRequestBody:
        title: DiaryUserSignupRequestBody
        type: object
        properties:
            email:
                type: string
                description: User email
                example: Impedit deserunt.
            name:
                type: string
                description: User name
                example: Modi est eos perspiciatis earum.
        example:
            email: Aut commodi dolorem eligendi dolor nemo provident.
            name: Voluptatem laudantium voluptatem est.
        required:
            - name
            - email
securityDefinitions:
    api_key_header_Authorization:
        type: apiKey
        description: Secures endpoint by requiring an API key.
        name: Authorization
        in: header